refactor(Squares): squares 분리
